Integrated sample kafka message with stock news rating model
