+++ requirements.txt
@@ -1,15 +1,13 @@
-pandas
-numpy
-networkx
-pyvis
-spacy
 matplotlib
+torch
+bs4
+pandas
+tqdm
+transformers
+numpy
+kafka-python
+
 Flask
 jupyter
-mpld3
 flask-cors
-scipy
-sklearn
-dill
-runipy
 pymysql
