+++ stable-req.txt
@@ -1,15 +1,13 @@
-pandas==1.0.1
-numpy==1.18.1
-networkx==2.4
-pyvis==0.1.7.0
-spacy==2.3.2
-matplotlib==3.1.3
-Flask==1.1.1
+matplotlib==3.3.3
+bs4==0.0.1
+pandas==1.2.1
+tqdm==4.56.0
+transformers==4.2.2
+numpy==1.19.5
+kafka-python==2.0.2
+torch==1.7.1
+
+Flask==1.1.2
 jupyter==1.0.0
-mpld3==0.3
-flask-cors==3.0.8
-scipy==1.4.1
-sklearn==0.0
-dill==0.3.1.1
-runipy==0.1.5
-pymysql==0.9.3
+flask-cors==3.0.10
+pymysql==1.0.2
